feat(add RTL expand file, and fix makefile)

+++ os/include/klib/string.h
@@ -5,9 +5,10 @@
                                                     Forrest Yu, 2005
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
 #include <kernel/const.h>
+#include <kernel/type.h>
 PUBLIC char *itoa(int num, char *str, int radix);
-PUBLIC void *memcpy(void *p_dst, const void *p_src, int size);
-PUBLIC void  memset(void *p_dst, char ch, int size);
+PUBLIC void *memcpy(void *p_dst, const void *p_src, size_t size);
+PUBLIC void  memset(void *p_dst, char ch, size_t size);
 PUBLIC char *strcpy(char *p_dst, const char *p_src);
 PUBLIC int   strcmp(const char *s1, const char *s2);
 // added by zcr
